fix(Runtime): Fix locomotion animation playback direction for embodied character
Summary:
# The problem
`AnimationHooks` uses a input locomotion vector to influence the character's animation tree. This means if vector's y-component is positive, the forward animation plays.

`MovementSDKLocomotion` provides this vector via the `OnUserInputChange` event, and the vector is also influenced by the PlayerController's orientation and the center eye's orientation. So if the user turns their body to the left without ratchet turn, the input vector's forward (z) direction will be slightly smaller. If they turn around completely, that forward direction will be negative. Notice how the `MovementSDKLocomotion`'s `OnUserInputChange` event is connected to the corresponding input field on `AnimationHooks`:

{F1657926971}

If the character turns their body such that the locomotion vector is negative, this causes the animation system to take a negative y for locomotion, causing the backwards animation to play. The same issue is described on [one of our GitHub issue](https://github.com/oculus-samples/Unity-Movement/issues/75) pages.

# The solution
The `OnUserInputChange` event can be modified to always match the user's current heading direction, but that might break expected behavior. This diff adds another field, `OnUserInputChangeJoystickDir` which always matches the user's joystick direction, and its scale is also modified by velocity. The scene's `AnimationHooks` was modified to use the `OnUserInputChangeJoystickDir` event instead:

 {F1657928491}

This fix is intended for the embodied character.

There was also some code that takes rotation differences between the avatar and camera orientation into account. The code was refactored into a separate function; the functionality remains the same.

Reviewed By: andkim-meta

Differential Revision: D57645058

fbshipit-source-id: 9513a4c8db77bd6cd64db3b15c62528631f8216e