feat(Runtime): Optimize FABRIK
Summary:
When used, the C# jobs of FABRIK seem to take 0.10-0.20 ms. It has to switch between reading transforms (which can be done efficiently in jobs) and the job form of the algorithm (which is serial in nature), before writing the final transforms. That seems to be slower than the plain version of FABRIK.

Since there is a possible overhead to creating jobs, it makes sense to use it for algorithms are more parallel in nature. FABRIK is more of a serial algorithm.

It makes more sense to clean up and optimize the code that runs FABRIK so that it uses poses instead of constantly accessing transform properties (which is slower).

Reviewed By: andkim-meta

Differential Revision: D58980107

fbshipit-source-id: 7316657b0d656df3b7365fec80b906cd24f3f63d