feat(Editor): Add functionality to update the current skeleton to the animator's avatar humanDescription
Summary: Add the "Update to Animator Pose" button to RetargetingLayer, which aligns the current skeleton with the rest pose of the animator avatar, using its skeleton data. This helps with the alignment calculations for retargeting, as the closer the retargeted character's rest pose matches with the rest pose of body tracking, the retargeting has less offsets to apply. In addition, Unity's humanoid avatar editor allows for changing an imported humanoid's rest pose into T-Pose, but that doesn't reflect on the model itself - this allows these T-Pose changes to be applied on the current character.

Reviewed By: sohailshafiiWk

Differential Revision: D59469818

fbshipit-source-id: d109d115eea905e894772ee1de6c004fe3e5c672