feat(Editor): Add JSON read/write for RetargetingLayer and RetargetingProcessor
Summary: Add JSON read/write serialization for RetargetingLayer and RetargetingProcessor. This allows serializing the adjustments and other important settings in RetargetingLayer to a JSON, as well as doing the same for any RetargetingProcessor. The inspector for these classes now display "Save to JSON" and "Read from JSON" buttons to make this process easier in Editor.

Reviewed By: sohailshafiiWk

Differential Revision: D59599900

fbshipit-source-id: 721cce0715a1213802077ba35ed7927f03b63541