fix(Runtime): Remove retargeting layer in AddMainRetargetingComponent
Summary:
`AddMainRetargetingComponent`, which adds retargeting layer, has confusing logic which gets the layer component and stores in two variables. It only removes the layer if one variable is null, and the other is not.

The code has been fixed to get the component once, and if it exists, it will be removed so that it can be replaced by a new component. This will accommodate the case where a character's mesh might change before `AddMainRetargetingComponent` is called, which will require a new retargeting layer to be added so that the retargeting offsets can be computed for the new mesh.

Reviewed By: andkim-meta

Differential Revision: D59607575

fbshipit-source-id: 934a5801bfdc2df8333a2ef486e2dac636317cbc