feat(Runtime): Add UpdateToAnimatorPose to AnimationUtilities
Summary: `UpdateToAnimatorPose` exists in `RetargetingLayer`. It's better to put it in `AnimationUtilities` because it should be used on avatars before `RetargetingLayer` is adding to them, if that component is added to them. This will allow a developer to align a character to T-pose before retargeting components are added to it.

Reviewed By: andkim-meta

Differential Revision: D59667116

fbshipit-source-id: 13daabe91c989acad08ba96ca0a6b42bac2ed4f3