fix(Runtime): Fix incorrect bone mappings being set when spawning an inactive character
Summary: When spawning a character and adding components to it through the AddComponentsHelper, the incorrect bone mapping could be assigned if the character has an inactive animator. This fixes that issue by updating the bone pair mapping after the character is active, and also adding an automatic bone pair mapping assignment if it hasn't occurred yet.

Reviewed By: sohailshafiiWk

Differential Revision: D59772494

fbshipit-source-id: 4ae3b316c7b3ba139a214b43a35d26ff7ecd9968