feat(Editor): Update one-click retargeting to take into account hand proportions for setting the finger position correction weight
Summary: Previously, the one-click retargeting option would always set the finger position correction weight to zero, since the hand proportions weren't taken into account when applying retargeting. This updates the function so that it will do a comparison between the rest pose hand and the model's hand to determine if they are within a size threshold to set the finger position correction weight.

Reviewed By: sohailshafiiWk

Differential Revision: D59974893

fbshipit-source-id: 2d2041eda27823e42c5e8ed427b7efbc80131c66