fix(Runtime): Fix character's rest pose before humanoid asset is used
Summary:
Previously, `SetupCharacterForAnimationRiggingRetargetingConstraints` would set the character to T-pose in the middle of the function. The problem is that a humanoid asset, used for adjustment calculation, was provided as an argument to the function and was computed on the character *before* it was forced to a T-pose. The creation of that asset will be affected by the character being in A-pose or T-pose since the humanoid asset used is different for each case.

This humanoid asset argument is used to create deformation's adjustments via `AddDeformationConstraint` below, however the line `AddComponentsHelper.AddJointAdjustments` ends up loading *another* humanoid asset on its, which could be different since the character has been snapped to T-pose at that point. Therefore, `AddDeformationConstraint` can use a A-pose humanoid asset if the character was in A-pose before, and then `AddComponentsHelper.AddJointAdjustments` will use the T-poes version.

This means the two sets of adjustments will differ, and this visually manifests as one-clicked characters end up have shoulders that droop. The character should be set to T-pose before the the setup function is called, so that all code paths will effectively load the same humanoid asset.

In addition, the call to `AddJointAdjustments` is provided the original humanoid asset passed in to `SetupCharacterForAnimationRiggingRetargetingConstraints` so that the code path uses the same humanoid asset.

Reviewed By: andkim-meta

Differential Revision: D60185986

fbshipit-source-id: 67f1ef0b24d73dfef644d987bde3e67fd20f0634