fix(samples): Fix for ISDK scene when controllers are used
Summary:
The ISDK scene was using a `OVRControllerHands` prefab for controller-driven hands. Its synthetic hands, which are ISDK hand meshes that represent how the hands wrap around the controllers, were being use to modify the body tracking hand poses. When they grabbed the cup and used plausible cup-holding poses, the ISDK-MSDK character's hands did the same. The integration was similar to the `OVRHands` prefab, which performed the same function but with the hands.

The problem is that `OVRControllerHands` is now deprecated and produces errors in the scene.

We can "Capsense" instead (https://developers.meta.com/horizon/documentation/unity/unity-capsense/), because that allows the hands to be driven even if controllers are being used. The changes are as follows:
1. Delete `OVRControllersHands` and also the references to its `SkeletalHandsAdjustment` component under `ArmatureSkinningUpdateRetargetSkeletonProcessor`. Delete the entries `LeftControllerHandSynthentic` and `RightControllerHandSynthentic`.
2. In `OVRCameraRig`->`OVRManager`, navigate to the "Controller Driven Hand Poses Type" and set it to Natural.
3. In each `OVRHandPrefab`, navigate to `OVRHand` and set "Show State" to Always (not something like "Contoroller Not In Hand").

Because `OVRControllerHands` is deleted and we can rely on `OVRHands` to be driven by controllers inputs, the scene is less deprecated and simpler.

Reviewed By: andkim-meta

Differential Revision: D63056976

fbshipit-source-id: edbf84648f52d1de6083b066960d48cd9f7c3ae9